feat: Randomize "Start Learning" button navigation
This commit changes the behavior of the "Start Learning" button on the home screen. Instead of always navigating to the phonetics screen, it now navigates to a random screen from the available menu items.

This is achieved by modifying the `onStartLearningClick` lambda in `HomeScreen.kt` to use `menuItems.random().route` for the navigation destination.