feat(phonetics): Improve `WavyProgressIndicator` styling
This commit enhances the visual styling of the `WavyProgressIndicator` used in the phonetics practice screen.

The line width and cap style have been updated to make the progress indicator thicker and have rounded ends. Additionally, the amplitude of the wave has been adjusted for a more pronounced visual effect.

The rest of the changes consist of code auto-formatting.